pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Analysis"
authors = [
    {name = "Nihar Dhanani", email = "nihardhanani28@gmail.com"},
]
description = "DSI building robust software assignment"
version = "0.1.0"
dependencies = ["matplotlib", "pyyaml", "requests", "pandas"]
requires-python = ">=3.10.0"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
Analysis = "Analysis:Analysis"